feat: final lesson 01

+++ modulo-2/aula-01/Curso.php
@@ -5,5 +5,5 @@ declare(strict_types=1);
 class Curso {
   public string $nome; 
   public string $descricao;
-  public int $duracao;
+  public float $duracao;
 }
